// src/app/components/weather/api/Enums.ts
export enum DaysInWeek {
    ALL = "all",
    MONDAY = "Monday",
    TUESDAY = "Tuesday",
    WEDNESDAY = "Wednesday",
    THURSDAY = "Thursday",
    FRIDAY = "Friday",
    SATURDAY = "Saturday",
    SUNDAY = "Sunday"
}

export enum WeatherStatus {
    SUNNY = "Sunny",
    CLOUDY = "Cloudy",
    RAINY = "Rainy",
}